Changement avec la connection et maj du tableau des fonctionnalité

+++ Programme/V6/PHP/src/vue/VueAffichageListe.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace mywishlist\vue;
 
 use mywishlist\models\Item;
+use mywishlist\models\Liste;
 
 
 class VueAffichageListe
@@ -35,18 +36,23 @@ class VueAffichageListe
 
         $_SESSION['titre'] = $l['titre'];
 
-        $items = Item::where('liste_id', '=', $l['no'])->get(); ;
+        $items = Item::where('liste_id', '=', $l['no'])->get();
+        $liste= Liste::where('no','=',$l['no'])->first();
 
         $html = "<h2>Liste {$l['no']}</h2>";
         $html .= "<b>Titre:</b> {$l['titre']}<br>";
         $html .= "<b>Description:</b> {$l['description']}<br>";
 
-        $html .= "<a href='$url_msg'>Ajouter un message à la liste</a><br>";
-        $html .= "<a href='$url_choixmodifyListe'>Modifier la liste</a><br>";
-        $html .= "<a href='$url_choixdeleteListe'>Supprimer la liste</a><br><br>";
+
+        if($liste['user_id']==$_SESSION['login']){
+            $html .= "<a href='$url_msg'>Ajouter un message à la liste</a><br>";
+            $html .= "<a href='$url_choixmodifyListe'>Modifier la liste</a><br>";
+            $html .= "<a href='$url_choixdeleteListe'>Supprimer la liste</a><br><br>";
+        }
 
         $html .= "<a href='$url_share'>Partager</a><br>";
         $html .= "<h3>Items </h3>";
+
 
         foreach($items as $item){
             $url_item   = $this->container->router->pathFor( 'aff_item', ['id' => $item['id']] ) ;
@@ -54,8 +60,9 @@ class VueAffichageListe
 
             $html .= "<li><a href='$url_item'>{$item['nom']}</a>,{$item['descr']}, {$item['tarif']}<a href='$url_reserv'><br><strong>RESERVER {$item['nom']}</strong></a></li><br>";
         }
-        $html .= "<a href='$url_form_item'>Ajouter un item</a><br><br>";
-
+        if($liste['user_id']==$_SESSION['login']) {
+            $html .= "<a href='$url_form_item'>Ajouter un item</a><br><br>";
+        }
         $html = "<ul>$html</ul>";
         return $html;
     }
